Backend-agnostic fixes from the Java branch, plus contract gate and cleanups (#4)
* Add automated Chase import pipeline

* Make revision ordering deterministic

Goal and journal revision listings ordered only by changed_at, which
ties within the same second and made revision order nondeterministic.
Add id DESC tie-breakers (ported from codex/java-backend-rewrite).


* Fix money field types and count invalidation in frontend

Endpoints without a response_model (monthly, series, recurring, goal
progress) serialize Decimal through jsonable_encoder as JSON numbers,
not strings; correct the TypeScript declarations to match. Also
invalidate transactions-count queries alongside transactions so counts
refresh after mutations (ported from codex/java-backend-rewrite).


* Remove unused pandas dependency

pandas is declared but imported nowhere in app/ or tests/; CSV parsing
uses the stdlib csv module and analytics use Decimal accumulation.
Dropping it removes pandas, numpy, and tzdata from the local lockfile.


* Attach response models to analytics and series routes

/api/analytics/monthly, /sankey, /recurring, and /api/snapshots/series
had no response_model, so their shapes were absent from the OpenAPI
schema and their wire format was whatever jsonable_encoder produced.
Attach the (previously dormant) MonthlyPoint/SankeyResponse/
RecurringMerchant models and a new NetWorthSeriesPoint, with money
typed as float to keep the historical JSON-number wire format —
Decimal fields would serialize as strings under Pydantic v2.

Verified byte-identical endpoint output before/after on a seeded
fixture, except untouched zero accumulators now serialize as 0.0
instead of 0 (numerically identical JSON; FastAPI's decimal_encoder
emitted whole Decimals as ints).

Goal progress is deliberately left schema-less: its early returns omit
as_of entirely, so any response model would change that wire shape.


* Cache rule description generalization

generate_rule_proposals evaluates every candidate key against every
transaction, and each evaluation re-generalizes the same descriptions
through ~a dozen regex passes — O(keys x transactions x regexes)
overall. The function is pure (str -> str), so an lru_cache collapses
that to one pass per distinct description: 41x faster on a simulated
50-key x 2000-transaction mining run.


* Share the CSV importer parse loop

The nine us_banks importers repeated the same read-enumerate-draft-
append loop, differing only in date/post-date keys, how a row yields
its amount and description, and the credit-card flag. Hoist the loop
into _parse_rows and express each importer as those per-row callables.
No behavior change; per-importer fixture tests all pass.


* Share a db fixture across backend tests

Every test file rolled its own in-memory engine + session + try/finally
close. Move that into a conftest.py db fixture and convert the ten
tests that used the pattern.


* Fix lint findings and consolidate NetWorthSeriesPoint

Chain HTTPException raises with 'from exc' (B904), apply import
sorting, and collapse the duplicate NetWorthSeriesPoint definition
flagged by F811 into the original net-worth-section model, now typed
float to match the series route's JSON-number wire format. ruff check
is fully clean.


* Freeze the API contract and gate it in CI

Snapshot the FastAPI OpenAPI document under docs/contracts/ and fail
the backend CI job when the committed snapshot is stale, so accidental
API changes surface in review instead of in the frontend. Refresh
intentionally with make api-contract-python; verify with
make api-contract-check. Ported from codex/java-backend-rewrite,
minus the Java-specific pieces.


---------

+++ backend/app/schemas.py
@@ -3,7 +3,8 @@
 # shadows the imported type once Pydantic re-resolves hints via
 # get_type_hints (which sees the class attribute set by the default value).
 # Aliasing the import sidesteps the shadow without renaming any JSON field.
-from datetime import date as _date, datetime
+from datetime import date as _date
+from datetime import datetime
 from decimal import Decimal
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -414,11 +415,15 @@ class NetWorthWorkbookImportResult(BaseModel):
 
 class NetWorthSeriesPoint(BaseModel):
     snapshot_date: _date
-    total: Decimal
-    by_category: dict[str, Decimal]
-    by_account: dict[str, Decimal]
-    taxable: Decimal
-    tax_advantaged: Decimal
+    # float, not Decimal — this route has always emitted JSON numbers
+    # (jsonable_encoder floats Decimals when no response_model is set), and
+    # the frontend charts consume numbers. Decimal fields would serialize as
+    # strings and change the wire format.
+    total: float
+    by_category: dict[str, float]
+    by_account: dict[str, float]
+    taxable: float
+    tax_advantaged: float
 
 
 # ---------- goals & journal ----------
@@ -580,15 +585,19 @@ class ImportBatchOut(ORMBase):
 
 
 class MonthlyPoint(BaseModel):
+    # float, not Decimal — this route has always emitted JSON numbers
+    # (jsonable_encoder floats Decimals when no response_model is set), and
+    # the frontend charts consume numbers. Decimal fields would serialize as
+    # strings and change the wire format.
     month: str  # YYYY-MM
-    by_kind: dict[str, Decimal]
-    by_expense_category: dict[str, Decimal]
-    by_income_category: dict[str, Decimal]
-    expenses_total: Decimal
-    income_total: Decimal
-    donations_total: Decimal
-    taxes_total: Decimal
-    net: Decimal
+    by_kind: dict[str, float]
+    by_expense_category: dict[str, float]
+    by_income_category: dict[str, float]
+    expenses_total: float
+    income_total: float
+    donations_total: float
+    taxes_total: float
+    net: float
 
 
 class ReconcileIn(BaseModel):
@@ -753,7 +762,8 @@ class SankeyResponse(BaseModel):
 class RecurringMerchant(BaseModel):
     merchant: str
     occurrences: int
-    avg_amount: Decimal
-    total_amount: Decimal
+    # float for the same wire-format reason as MonthlyPoint.
+    avg_amount: float
+    total_amount: float
     last_seen: _date
     cadence_days_estimate: float | None
